+++ CHANGELOG.md
@@ -4,12 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2023-03-03
+## [Unreleased] - 2023-03-05
 
 ### Added
 
 ### Changed
-* Minor code improvements within Permutation and PermutationIterator classes.
+* Minor code improvements to Permutation and PermutationIterator classes.
+* Minor code improvements to KendallTauDistance and WeightedKendallTauDistance.
+* Minor code improvements to CyclicEdgeDistance and CyclicRTypeDistance.
 
 ### Deprecated